feat(web/homepage): Port flurry, a star torn apart into coloured smoke
Three kinds of thing, only one of which you can see. The star is a point
on a Lissajous figure put through rotations at speeds that are not
multiples of each other, so it never repeats, and a puff of smoke leaves
it per stream every hundred and twenty-first of a second. The sparks are
the attractors and are never drawn: upstream's code to draw them is
behind an ifdef nothing sets. Each smoke particle is pulled towards
every spark, with an extra helping for the one whose number matches its
own, which is what keeps the streams apart instead of letting them
average into a blob.

Each puff is a screen-space quad stretched along the way it moved this
frame, so a fast particle is a streak and a slow one a dot, and it fades
as it widens until it is wider than the stream and dies. The texture is
an eight by eight sheet of soft blobs made at startup, each a cosine
falloff speckled with noise and smoothed twice.

Nothing is ever cleared: a black rectangle is laid over the screen every
frame and the smoke added on top. A colour buffer does not survive to
the next frame here, so the screen is kept in a texture and put back at
the start of one, the way noof does it.

Two things are done differently. Upstream reads the wall clock and
refuses to draw a frame that comes sooner than a sixtieth of a second
after the last; here the host asks for a frame and is told how long to
wait before the next, so the clock is taken from the delay and the two
stay in step. And what upstream calls the frame rate is its frame count
over the time since it started plus a random offset of up to five
minutes, so for the first minutes of a run everything pulls several
times too hard and the flurry is a few streaks rather than a bloom. The
offset is taken back out, which makes the quantity the frame rate it is
named after from the first frame.

+++ web/homepage/xscreensaver/src/runtime/quickhull.rs
@@ -968,9 +968,9 @@ mod tests {
         let corners = points.clone();
         for _ in 0..400 {
             points.push(Vertex::new(
-                (frand(1.6) - 0.8) as f64,
-                (frand(1.6) - 0.8) as f64,
-                (frand(1.6) - 0.8) as f64,
+                frand(1.6) - 0.8,
+                frand(1.6) - 0.8,
+                frand(1.6) - 0.8,
             ));
         }
         let hull = quickhull3d(&points);
@@ -991,11 +991,7 @@ mod tests {
         ya_rand_init(20260812);
         let mut points = Vec::new();
         while points.len() < 2000 {
-            let v = Vertex::new(
-                (0.5 - frand(1.0)) as f64,
-                (0.5 - frand(1.0)) as f64,
-                (0.5 - frand(1.0)) as f64,
-            );
+            let v = Vertex::new(0.5 - frand(1.0), 0.5 - frand(1.0), 0.5 - frand(1.0));
             if v.length2() < 0.25 {
                 points.push(v);
             }
